requirements.txt
requests
python-dotenv
pandas
boto3 # to interact with MinIO from Python
websockets
minio
dbt-core
dbt-postgres
# dbt-duckdb
# duckdb
pandas
# superset requirements were passed thru docker compose bash command
sqlalchemy
psycopg2-binary
great_expectations # not used yet